fall24/AppSoftwareDev/wsu-wams-ng-main/src/app/shared/modal/modal.component.ts
import {Component, EventEmitter, HostListener, Input, Output} from '@angular/core';
import {NgIf} from "@angular/common";

@Component({
  selector: 'app-modal',
  standalone: true,
  imports: [
    NgIf
  ],
  templateUrl: './modal.component.html',
  styleUrl: './modal.component.scss'
})
export class ModalComponent {
  @Input() show = false;
  @Output() close = new EventEmitter<void>();

  @HostListener('document:keydown.escape', ['$event'])
  onEscKey(event: KeyboardEvent) {
    if (this.show) {
      this.onClose();
    }
  }

  onClose() {
    this.close.emit();
  }
}
